requirements.txt
# requirements.txt
# ── تثبيت الكل: pip install -r requirements.txt ────────────────
greenlet==3.0.3

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0          # PostgreSQL async driver
alembic==1.13.3          # Database migrations

# AI
anthropic==0.34.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# Utils
python-dotenv==1.0.1
httpx==0.27.2            # للطلبات HTTP

# PDF Generation (ورقة العمل)
reportlab==4.2.5

# Data Analysis (مستقبلاً)
# pandas==2.2.0
# scikit-learn==1.5.0
# numpy==2.1.0
